requirements.txt
streamlit
langchain
sentence-transformers
pypdf2
psycopg2-binary
python-dotenv
pgvector